Support multiple spaces in in log_format.patterns
Fixes #7

We currently do not support multiple spaces between the format glob and
the format name. This PR fixes that by considering consecutive
whitespace as a single separator.

I also added documentation to the configuration file to make it clear
that at least 1 leading whitespace character is required to signify that
a line is a continuation of the previous.